+++ api/rating.go
@@ -23,7 +23,7 @@ type Opponent struct {
 }
 
 func (o Opponent) R() float64 {
-	return o.Rating
+	return float64(o.Rating)
 }
 
 func (o Opponent) RD() float64 {
@@ -51,7 +51,7 @@ func newOpponent(p Player, outcome result) Opponent {
 // TODO: this works on a per-game basis. The glicko2 doc says it works best with a
 // longer rating period of 10-15 games, so maybe this function should take a single
 // Player, look up their previous ~10 games and calculate the rating based on that.
-func rank(ctx context.Context, g *Game) (float64, float64) {
+func rank(ctx context.Context, g *Game) (int, int) {
 	winner, loser := g.WinnerAndLoser()
 
 	// winnerOpponent is the loser (the opponent of the winner)
@@ -67,18 +67,19 @@ func rank(ctx context.Context, g *Game) (float64, float64) {
 	return winner.Rating, loser.Rating
 }
 
-func rankPlayer(p *Player, opponents []glicko2.Opponent) float64 {
+// rankPlayer updates this Player's rank and returns the difference
+func rankPlayer(p *Player, opponents []glicko2.Opponent) int {
 	prev := p.Rating
 
 	r, rd, rv := glicko2.Rank(
-		p.Rating,
+		float64(p.Rating),
 		p.RatingDeviation,
 		p.RatingVolatility,
 		opponents,
 		τ,
 	)
 
-	p.Rating, p.RatingDeviation, p.RatingVolatility = r, rd, rv
+	p.Rating, p.RatingDeviation, p.RatingVolatility = int(r), rd, rv
 
 	return p.Rating - prev
 }
